pyproject.toml
[project]
name = "fzuhelper-bot"
version = "0.1.1"
description = "fzuhelper-bot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "aiohttp>=3.13.4",
    "aiosmtplib>=5.1.0",
    "nb-cli>=1.7.2",
    "nonebot-adapter-milky>=1.1.0",
    "nonebot2[fastapi,httpx]>=2.4.4",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []
[tool.nonebot.plugins]
"@local" = []


[tool.nonebot.adapters]
"@local" = []
nonebot-adapter-milky = [{ name = "nonebot-adapter-milky", module_name = "nonebot.adapters.milky" }]
